refactor(cmd): Migrate all `kraft` commands into `internal/cli/kraft`
