+++ internal/cli/kraft/cloud/img/list/list.go
