+++ internal/cli/kraft/cloud/instance/remove/remove.go
