+++ internal/cli/kraft/cloud/instance/status/status.go
@@ -2,7 +2,8 @@
 // Copyright (c) 2023, Unikraft GmbH and The KraftKit Authors.
 // Licensed under the BSD-3-Clause License (the "License").
 // You may not use this file except in compliance with the License.
-package list
+
+package status
 
 import (
 	"fmt"
@@ -14,27 +15,28 @@ import (
 	kraftcloud "sdk.kraft.cloud"
 	kcinstance "sdk.kraft.cloud/instance"
 
-	"kraftkit.sh/cmd/kraft/cloud/utils"
 	"kraftkit.sh/cmdfactory"
 	"kraftkit.sh/config"
+	"kraftkit.sh/internal/cli/kraft/cloud/utils"
 	"kraftkit.sh/log"
 )
 
-type List struct {
+type Status struct {
 	Output string `long:"output" short:"o" usage:"Set output format" default:"table"`
 
 	metro string
 }
 
 func New() *cobra.Command {
-	cmd, err := cmdfactory.New(&List{}, cobra.Command{
-		Short:   "List instances",
-		Use:     "ls [FLAGS]",
-		Aliases: []string{"list"},
+	cmd, err := cmdfactory.New(&Status{}, cobra.Command{
+		Short:   "Retrieve the status of an instance",
+		Use:     "status [FLAGS] UUID",
+		Args:    cobra.ExactArgs(1),
+		Aliases: []string{"info"},
 		Example: heredoc.Doc(`
-			# List all instances in your account.
-			$ kraft cloud instances list
-		`),
+			# Retrieve information about a kraftcloud instance
+			$ kraft cloud instance status fd1684ea-7970-4994-92d6-61dcc7905f2b
+	`),
 		Annotations: map[string]string{
 			cmdfactory.AnnotationHelpGroup: "kraftcloud-instance",
 		},
@@ -46,7 +48,7 @@ func New() *cobra.Command {
 	return cmd
 }
 
-func (opts *List) Pre(cmd *cobra.Command, _ []string) error {
+func (opts *Status) Pre(cmd *cobra.Command, _ []string) error {
 	opts.metro = cmd.Flag("metro").Value.String()
 	if opts.metro == "" {
 		opts.metro = os.Getenv("KRAFTCLOUD_METRO")
@@ -58,7 +60,7 @@ func (opts *List) Pre(cmd *cobra.Command, _ []string) error {
 	return nil
 }
 
-func (opts *List) Run(cmd *cobra.Command, args []string) error {
+func (opts *Status) Run(cmd *cobra.Command, args []string) error {
 	ctx := cmd.Context()
 	auth, err := config.GetKraftCloudLoginFromContext(ctx)
 	if err != nil {
@@ -69,23 +71,10 @@ func (opts *List) Run(cmd *cobra.Command, args []string) error {
 		kraftcloud.WithToken(auth.Token),
 	)
 
-	uuids, err := client.WithMetro(opts.metro).List(ctx)
+	instance, err := client.WithMetro(opts.metro).Status(ctx, args[0])
 	if err != nil {
-		return fmt.Errorf("could not list instances: %w", err)
+		return fmt.Errorf("could not create instance: %w", err)
 	}
 
-	// TODO(nderjung): For now, the KraftCloud API does not support
-	// returning the full details of each instance.  Temporarily request a
-	// status for each instance.
-	var instances []kcinstance.Instance
-	for _, uuid := range uuids {
-		instance, err := client.WithMetro(opts.metro).Status(ctx, uuid.UUID)
-		if err != nil {
-			return fmt.Errorf("could not get instance status: %w", err)
-		}
-
-		instances = append(instances, *instance)
-	}
-
-	return utils.PrintInstances(ctx, opts.Output, instances...)
+	return utils.PrintInstances(ctx, opts.Output, *instance)
 }
