+++ internal/cli/kraft/cloud/instance/stop/stop.go
