+++ internal/cli/kraft/net/inspect/inspect.go
