+++ internal/cli/kraft/net/list/list.go
