+++ internal/cli/kraft/net/remove/remove.go
