+++ internal/cli/kraft/pkg/rm/rm.go
