+++ internal/cli/kraft/properclean/properclean.go
