+++ internal/cli/kraft/stop/stop.go
